requirements.txt
aniso8601==7.0.0
backpack==0.1
bcrypt==4.0.1
blinker==1.5
cleo==0.6.8
click==7.1.2
contextlib2==21.6.0
ecdsa==0.18.0
Faker==0.8.18
fastapi==0.61.1
graphene==2.1.8
graphene-pydantic==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.14.0
inflection==0.3.1
lazy-object-proxy==1.8.0
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
promise==2.3
psycopg2-binary==2.8.6
pyaml==16.12.2
pyasn1==0.4.8
pydantic==1.10.2
Pygments==2.13.0
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.7
pytzdata==2020.1
PyYAML==5.4.1
rsa==4.9
Rx==1.6.3
schema==0.7.5
simplejson==3.18.0
six==1.16.0
starlette==0.13.6
text-unidecode==1.2
typing-extensions==4.4.0
tzlocal==1.5.1
uvicorn==0.12.2
wrapt==1.14.1
